backend/requirements.txt
# Django y dependencias principales
Django>=5.0,<6.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
djangorestframework-simplejwt>=5.0.0

# Base de datos
psycopg2-binary>=2.9.0
dj-database-url>=2.0.0

# Variables de entorno
python-dotenv>=1.0.0

# Archivos estáticos y media
whitenoise>=6.0.0
Pillow>=10.0.0

# Utilidades básicas
requests>=2.30.0

# Producción
gunicorn>=21.0.0
